Catch and log exceptions in edge callback. Abort program.
It turns out that I was getting an exception in my callback function,
but I didn't know because (a) lidet.py didn't exit when an exception
happened inside the callback, and (b) there was no visible output
to notice the exception stack trace being printed.

So now when an exception happens in the callback, log it to logger.log
and set a flag that causes the while program to exit with code 8.